User now cannot delete himself Possiblly fixed #32 Fixed issue where no new receipt is made when you start searching for items. Added new panel for only updating items instead of importing them all (WIP)

+++ src/management/deleteUser.php
@@ -3,17 +3,24 @@ include_once("../includes.php");
 
 if ($_GET['userId'] != "")
 {
-    if (Misc::sqlExists("userId", $_GET['userId'], "users"))
+    if ($_GET['userId'] != $_SESSION['login_ok']['userId'])
     {
-        $result = Misc::sql('DELETE FROM users WHERE userId=' . $_GET['userId'] . ';');
+        if (Misc::sqlExists("userId", $_GET['userId'], "users"))
+        {
+            $result = Misc::sql('DELETE FROM users WHERE userId=' . $_GET['userId'] . ';');
 
-        if ($result == 1)
-            die('OK');
+            if ($result == 1)
+                die('OK');
+            else
+                die('Er is iets fout gegaan tijdens het verwijderen van de gebruiker.');
+        }
         else
-            die('Er is iets fout gegaan tijdens het verwijderen van de gebruiker.');
+        {
+            die('Deze gebruiker bestaat niet!');
+        }
     }
     else
     {
-        die('Deze gebruiker bestaat niet!');
+        die('Je kunt deze gebruiker niet verwijderen. Je probeert jezelf te verwijderen ?!');
     }
 }
